Handle BIOS and UEFI in one go
We need to figure out if the system is UEFI or BIOS
and based on this different partitions and commands to install grub
are used.

Also start on a configure script for after install configuration.
Since now we only have the root user, we need to add a new user
which is a TODO for now.